+++ rust/src/shapes/request.rs
@@ -503,11 +503,6 @@ pub struct RequestParams {
     /// Evaluates given script in every frame upon creation (before loading frame's scripts).
     pub evaluate_on_new_document: Option<Box<String>>,
     #[serde(default)]
-    /// Runs the request using lite_mode:Lite mode reduces data transfer costs by 50%, with trade-offs in speed, accuracy,
-    /// geo-targeting, and reliability. It’s best suited for non-urgent data collection or when
-    /// targeting websites with minimal anti-bot protections.
-    pub lite_mode: Option<bool>,
-    #[serde(default)]
     /// The proxy to use for request.
     pub proxy: Option<ProxyType>,
     #[serde(default)]